trashnet/requirements.txt
tensorflow
numpy
flask
matplotlib
pillow